fix: Change of luminance value calculation and scene detection process
The following parameters are specified to detect black scenes.
In the timestamp generation process, the frame luminance value was compared with the frame luminance value and saved in the list if it matched the parameter.
However, there were cases where the luminance values were of float type and did not match, so this has been corrected.

seane_detect_enable=true
seane_detect_mean_val = 0